Refactored the project into several packages
