+++ Week 6/MovieAppW6/app/src/main/java/com/janewaitara/movieapp/MainActivity.kt
@@ -1,11 +1,10 @@
 package com.janewaitara.movieapp
 
 import android.os.Bundle
-import android.view.Menu
-import android.view.MenuItem
 import androidx.appcompat.app.AppCompatActivity
 import androidx.navigation.NavController
 import androidx.navigation.Navigation
+import com.janewaitara.movieapp.storage.MovieSharedPrefs
 
 class MainActivity : AppCompatActivity() {
