+++ Week 6/MovieAppW6/app/src/main/java/com/janewaitara/movieapp/model/Movie.kt
@@ -1,4 +1,4 @@
-package com.janewaitara.movieapp
+package com.janewaitara.movieapp.model
 
 import android.os.Parcel
 import android.os.Parcelable
