+++ Week 6/MovieAppW6/app/src/main/java/com/janewaitara/movieapp/ui/movies/MovieListFragment.kt
@@ -1,4 +1,4 @@
-package com.janewaitara.movieapp
+package com.janewaitara.movieapp.ui.movies
 
 import android.os.Bundle
 import android.view.*
@@ -6,10 +6,13 @@ import androidx.fragment.app.Fragment
 import androidx.lifecycle.Observer
 import androidx.lifecycle.ViewModelProvider
 import androidx.lifecycle.lifecycleScope
-import androidx.navigation.NavController
 import androidx.navigation.findNavController
 import androidx.recyclerview.widget.RecyclerView
 import androidx.recyclerview.widget.StaggeredGridLayoutManager
+import com.janewaitara.movieapp.MovieListFragmentDirections
+import com.janewaitara.movieapp.storage.MovieSharedPrefs
+import com.janewaitara.movieapp.R
+import com.janewaitara.movieapp.model.Movie
 import kotlinx.coroutines.launch
 
 class MovieListFragment : Fragment(), MovieAdapter.MovieListClickListener {
@@ -53,7 +56,10 @@ class MovieListFragment : Fragment(), MovieAdapter.MovieListClickListener {
 
     private fun showDetailsActivity(movie: Movie){
         view?.let {
-            val action = MovieListFragmentDirections.actionMovieListFragmentToMovieDetailFragment(movie) //trigger the navigation and passing data
+            val action =
+                MovieListFragmentDirections.actionMovieListFragmentToMovieDetailFragment(
+                    movie
+                ) //trigger the navigation and passing data
             it.findNavController().navigate(action)
         }
