+++ Week 6/MovieAppW6/app/src/main/res/layout/fragment_movie_list.xml
@@ -4,7 +4,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     xmlns:app="http://schemas.android.com/apk/res-auto"
-    tools:context=".MovieListFragment">
+    tools:context=".ui.movies.MovieListFragment">
 
     <androidx.recyclerview.widget.RecyclerView
         android:id="@+id/movieRecyclerView"
